fix(ffe): share one request-scoped span-enrichment accumulator across all paths
PR review (DataDog/dd-trace-php#3996) blocker + should-fix.

blocker: tracer/ffe.c set_ffe_span_enrichment_tags() REPLACES the three
request-global tag slots on every call. Both DataDogProvider and each
FeatureFlags\Client/SpanEnrichmentBinder owned a SEPARATE accumulator and
staged independently, so two clients, two providers, or a mixed OpenFeature +
native-client evaluation under ONE root span would OVERWRITE earlier serial
ids / hashed subjects / runtime defaults instead of aggregating them.

Fix: introduce SpanEnrichmentRegistry, a single request-scoped accumulator
that ALL PHP evaluation paths feed. The staged tag set is now the union of
every evaluation on the active root span, matching the frozen Node contract.
No tag/encoding/limit semantics changed.

should-fix (per-binder onClose retention): the lifecycle is centralized in the
registry, which binds AT MOST ONE root-close reset per root span (tracked by
rootCloseBoundRootId). Many short-lived clients under one long-lived root no
longer each retain a closure + accumulator. SpanEnrichmentBinder is now a thin
gate-checked adapter; DataDogProvider drops its inline accumulator + lifecycle.

should-fix (gate-off not inert): Client and DataDogProvider now construct NO
binder unless DD_EXPERIMENTAL_FLAGGING_PROVIDER_SPAN_ENRICHMENT_ENABLED is on,
and evaluate()/resolve() skip the enrichment call entirely when the binder is
absent — no per-evaluation config read with the gate off (DG-005).

should-fix (root side effect): the registry resolves the root id via the new
non-creating DDTrace\Internal\peek_root_span_id(), falling back to the
(creating) DDTrace\root_span() only on older extensions.
